+++ src/network.rs
@@ -1,0 +1,1 @@
+pub mod p2p;
